requirements.txt
# Core dependencies
qdrant-client==1.6.4
sentence-transformers==2.2.2
langchain==0.0.335
langchain-community==0.0.10
langchainhub==0.1.14
torch>=2.0.0
transformers>=4.34.0
openai>=1.3.0

# Document processing
pypdf==3.17.1
python-docx==0.8.11
docx2txt==0.8
markdown==3.5
beautifulsoup4==4.12.2

# API and services
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
numpy==1.26.0

# Testing
pytest==7.4.3